re-consolidate tests ...
... use rstest to run same tests for standalone
and remote.
