+++ ballista/client/Cargo.toml
@@ -44,12 +44,11 @@ url = { workspace = true }
 [dev-dependencies]
 ballista-executor = { path = "../executor", version = "0.12.0" }
 ballista-scheduler = { path = "../scheduler", version = "0.12.0" }
-ctor = { version = "0.2" }
+ctor = { workspace = true }
 env_logger = { workspace = true }
-object_store = { workspace = true, features = ["aws"] }
-testcontainers-modules = { version = "0.11", features = ["minio"] }
+rstest = { version = "0.23" }
+tonic = { workspace = true }
 
 [features]
 default = ["standalone"]
 standalone = ["ballista-executor", "ballista-scheduler"]
-testcontainers = []
