+++ examples/Cargo.toml
@@ -52,3 +52,13 @@ tokio = { workspace = true, features = [
     "parking_lot"
 ] }
 url = { workspace = true }
+
+[dev-dependencies]
+ctor = { workspace = true }
+env_logger = { workspace = true }
+testcontainers-modules = { version = "0.11", features = ["minio"] }
+tonic = { workspace = true }
+
+[features]
+default = []
+testcontainers = []
