+++ examples/tests/object_store.rs
@@ -27,11 +27,11 @@
 mod common;
 
 #[cfg(test)]
-#[cfg(feature = "standalone")]
 #[cfg(feature = "testcontainers")]
 mod standalone {
 
     use ballista::extension::SessionContextExt;
+    use ballista_examples::test_util::examples_test_data;
     use datafusion::{assert_batches_eq, prelude::SessionContext};
     use datafusion::{
         error::DataFusionError,
@@ -52,12 +52,13 @@ mod standalone {
             .await
             .unwrap();
 
+        let host = node.get_host().await.unwrap();
         let port = node.get_host_port_ipv4(9000).await.unwrap();
 
-        let object_store = crate::common::create_s3_store(port)
+        let object_store = crate::common::create_s3_store(&host.to_string(), port)
             .map_err(|e| DataFusionError::External(e.into()))?;
 
-        let test_data = crate::common::example_test_data();
+        let test_data = examples_test_data();
         let config = RuntimeConfig::new();
         let runtime_env = RuntimeEnv::new(config)?;
 
@@ -116,6 +117,7 @@ mod standalone {
 mod remote {
 
     use ballista::extension::SessionContextExt;
+    use ballista_examples::test_util::examples_test_data;
     use datafusion::{assert_batches_eq, prelude::SessionContext};
     use datafusion::{
         error::DataFusionError,
@@ -129,7 +131,7 @@ mod remote {
 
     #[tokio::test]
     async fn should_execute_sql_write() -> datafusion::error::Result<()> {
-        let test_data = crate::common::example_test_data();
+        let test_data = examples_test_data();
 
         let container = crate::common::create_minio_container();
         let node = container.start().await.unwrap();
@@ -138,9 +140,10 @@ mod remote {
             .await
             .unwrap();
 
+        let host = node.get_host().await.unwrap();
         let port = node.get_host_port_ipv4(9000).await.unwrap();
 
-        let object_store = crate::common::create_s3_store(port)
+        let object_store = crate::common::create_s3_store(&host.to_string(), port)
             .map_err(|e| DataFusionError::External(e.into()))?;
 
         let config = RuntimeConfig::new();
@@ -210,15 +213,12 @@ mod remote {
 #[cfg(feature = "testcontainers")]
 mod custom_s3_config {
 
+    use crate::common::{ACCESS_KEY_ID, SECRET_KEY};
     use ballista::extension::SessionContextExt;
     use ballista::prelude::SessionConfigExt;
     use ballista_core::RuntimeProducer;
-    use datafusion::common::{config_err, exec_err};
-    use datafusion::config::{
-        ConfigEntry, ConfigExtension, ConfigField, ExtensionOptions, Visit,
-    };
-    use datafusion::error::Result;
-    use datafusion::execution::object_store::ObjectStoreRegistry;
+    use ballista_examples::object_store::{CustomObjectStoreRegistry, S3Options};
+    use ballista_examples::test_util::examples_test_data;
     use datafusion::execution::SessionState;
     use datafusion::prelude::SessionConfig;
     use datafusion::{assert_batches_eq, prelude::SessionContext};
@@ -229,22 +229,13 @@ mod custom_s3_config {
             SessionStateBuilder,
         },
     };
-    use object_store::aws::AmazonS3Builder;
-    use object_store::local::LocalFileSystem;
-    use object_store::ObjectStore;
-    use parking_lot::RwLock;
-    use std::any::Any;
-    use std::fmt::Display;
     use std::sync::Arc;
     use testcontainers_modules::testcontainers::runners::AsyncRunner;
-    use url::Url;
-
-    use crate::common::{ACCESS_KEY_ID, SECRET_KEY};
 
     #[tokio::test]
     async fn should_configure_s3_execute_sql_write_remote(
     ) -> datafusion::error::Result<()> {
-        let test_data = crate::common::example_test_data();
+        let test_data = examples_test_data();
 
         //
         // Minio cluster setup
@@ -256,7 +247,14 @@ mod custom_s3_config {
             .await
             .unwrap();
 
+        let endpoint_host = node.get_host().await.unwrap();
         let endpoint_port = node.get_host_port_ipv4(9000).await.unwrap();
+
+        log::info!(
+            "MINIO testcontainers host: {}, port: {}",
+            endpoint_host,
+            endpoint_port
+        );
 
         //
         // Session Context and Ballista cluster setup
@@ -325,8 +323,8 @@ mod custom_s3_config {
             .show()
             .await?;
         ctx.sql(&format!(
-            "SET s3.endpoint = 'http://localhost:{}'",
-            endpoint_port
+            "SET s3.endpoint = 'http://{}:{}'",
+            endpoint_host, endpoint_port
         ))
         .await?
         .show()
@@ -380,10 +378,9 @@ mod custom_s3_config {
     // SessionConfig propagation across ballista cluster.
 
     #[tokio::test]
-    #[cfg(feature = "standalone")]
     async fn should_configure_s3_execute_sql_write_standalone(
     ) -> datafusion::error::Result<()> {
-        let test_data = crate::common::example_test_data();
+        let test_data = examples_test_data();
 
         //
         // Minio cluster setup
@@ -419,11 +416,6 @@ mod custom_s3_config {
 
         let session_builder = Arc::new(produce_state);
         let state = session_builder(config_producer());
-
-        // // setting up ballista cluster with new runtime, configuration, and session state producers
-        // let (host, port) =
-        //     crate::common::setup_test_cluster_with_state(state.clone()).await;
-        // let url = format!("df://{host}:{port}");
 
         // // establishing cluster connection,
         let ctx: SessionContext = SessionContext::standalone_with_state(state).await?;
@@ -506,232 +498,5 @@ mod custom_s3_config {
             .with_runtime_env(runtime_env.into())
             .with_config(session_config)
             .build()
-    }
-
-    #[derive(Debug)]
-    pub struct CustomObjectStoreRegistry {
-        local: Arc<LocalFileSystem>,
-        s3options: S3Options,
-    }
-
-    impl CustomObjectStoreRegistry {
-        fn new(s3options: S3Options) -> Self {
-            Self {
-                s3options,
-                local: Arc::new(LocalFileSystem::new()),
-            }
-        }
-    }
-
-    impl ObjectStoreRegistry for CustomObjectStoreRegistry {
-        fn register_store(
-            &self,
-            _url: &Url,
-            _store: Arc<dyn ObjectStore>,
-        ) -> Option<Arc<dyn ObjectStore>> {
-            unreachable!("register_store not supported ")
-        }
-
-        fn get_store(&self, url: &Url) -> Result<Arc<dyn ObjectStore>> {
-            let scheme = url.scheme();
-            log::info!("get_store: {:?}", &self.s3options.config.read());
-            match scheme {
-                "" | "file" => Ok(self.local.clone()),
-                "s3" => {
-                    let s3store = Self::s3_object_store_builder(
-                        url,
-                        &self.s3options.config.read(),
-                    )?
-                    .build()?;
-
-                    Ok(Arc::new(s3store))
-                }
-
-                _ => exec_err!("get_store - store not supported, url {}", url),
-            }
-        }
-    }
-
-    impl CustomObjectStoreRegistry {
-        pub fn s3_object_store_builder(
-            url: &Url,
-            aws_options: &S3RegistryConfiguration,
-        ) -> Result<AmazonS3Builder> {
-            let S3RegistryConfiguration {
-                access_key_id,
-                secret_access_key,
-                session_token,
-                region,
-                endpoint,
-                allow_http,
-            } = aws_options;
-
-            let bucket_name = Self::get_bucket_name(url)?;
-            let mut builder = AmazonS3Builder::from_env().with_bucket_name(bucket_name);
-
-            if let (Some(access_key_id), Some(secret_access_key)) =
-                (access_key_id, secret_access_key)
-            {
-                builder = builder
-                    .with_access_key_id(access_key_id)
-                    .with_secret_access_key(secret_access_key);
-
-                if let Some(session_token) = session_token {
-                    builder = builder.with_token(session_token);
-                }
-            } else {
-                return config_err!(
-                    "'s3.access_key_id' & 's3.secret_access_key' must be configured"
-                );
-            }
-
-            if let Some(region) = region {
-                builder = builder.with_region(region);
-            }
-
-            if let Some(endpoint) = endpoint {
-                if let Ok(endpoint_url) = Url::try_from(endpoint.as_str()) {
-                    if !matches!(allow_http, Some(true))
-                        && endpoint_url.scheme() == "http"
-                    {
-                        return config_err!("Invalid endpoint: {endpoint}. HTTP is not allowed for S3 endpoints. To allow HTTP, set 's3.allow_http' to true");
-                    }
-                }
-
-                builder = builder.with_endpoint(endpoint);
-            }
-
-            if let Some(allow_http) = allow_http {
-                builder = builder.with_allow_http(*allow_http);
-            }
-
-            Ok(builder)
-        }
-
-        fn get_bucket_name(url: &Url) -> Result<&str> {
-            url.host_str().ok_or_else(|| {
-                DataFusionError::Execution(format!(
-                    "Not able to parse bucket name from url: {}",
-                    url.as_str()
-                ))
-            })
-        }
-    }
-
-    #[derive(Debug, Clone, Default)]
-    pub struct S3Options {
-        config: Arc<RwLock<S3RegistryConfiguration>>,
-    }
-
-    impl ExtensionOptions for S3Options {
-        fn as_any(&self) -> &dyn Any {
-            self
-        }
-
-        fn as_any_mut(&mut self) -> &mut dyn Any {
-            self
-        }
-
-        fn cloned(&self) -> Box<dyn ExtensionOptions> {
-            Box::new(self.clone())
-        }
-
-        fn set(&mut self, key: &str, value: &str) -> Result<()> {
-            log::debug!("set config, key:{},  value:{}", key, value);
-            match key {
-                "access_key_id" => {
-                    let mut c = self.config.write();
-                    c.access_key_id.set(key, value)?;
-                }
-                "secret_access_key" => {
-                    let mut c = self.config.write();
-                    c.secret_access_key.set(key, value)?;
-                }
-                "session_token" => {
-                    let mut c = self.config.write();
-                    c.session_token.set(key, value)?;
-                }
-                "region" => {
-                    let mut c = self.config.write();
-                    c.region.set(key, value)?;
-                }
-                "endpoint" => {
-                    let mut c = self.config.write();
-                    c.endpoint.set(key, value)?;
-                }
-                "allow_http" => {
-                    let mut c = self.config.write();
-                    c.allow_http.set(key, value)?;
-                }
-                _ => {
-                    log::warn!("Config value {} cant be set to {}", key, value);
-                    return config_err!(
-                        "Config value \"{}\" not found in S3Options",
-                        key
-                    );
-                }
-            }
-            Ok(())
-        }
-
-        fn entries(&self) -> Vec<ConfigEntry> {
-            struct Visitor(Vec<ConfigEntry>);
-
-            impl Visit for Visitor {
-                fn some<V: Display>(
-                    &mut self,
-                    key: &str,
-                    value: V,
-                    description: &'static str,
-                ) {
-                    self.0.push(ConfigEntry {
-                        key: format!("{}.{}", S3Options::PREFIX, key),
-                        value: Some(value.to_string()),
-                        description,
-                    })
-                }
-
-                fn none(&mut self, key: &str, description: &'static str) {
-                    self.0.push(ConfigEntry {
-                        key: format!("{}.{}", S3Options::PREFIX, key),
-                        value: None,
-                        description,
-                    })
-                }
-            }
-            let c = self.config.read();
-
-            let mut v = Visitor(vec![]);
-            c.access_key_id
-                .visit(&mut v, "access_key_id", "S3 Access Key");
-            c.secret_access_key
-                .visit(&mut v, "secret_access_key", "S3 Secret Key");
-            c.session_token
-                .visit(&mut v, "session_token", "S3 Session token");
-            c.region.visit(&mut v, "region", "S3 region");
-            c.endpoint.visit(&mut v, "endpoint", "S3 Endpoint");
-            c.allow_http.visit(&mut v, "allow_http", "S3 Allow Http");
-
-            v.0
-        }
-    }
-
-    impl ConfigExtension for S3Options {
-        const PREFIX: &'static str = "s3";
-    }
-    #[derive(Default, Debug, Clone)]
-    pub struct S3RegistryConfiguration {
-        /// Access Key ID
-        pub access_key_id: Option<String>,
-        /// Secret Access Key
-        pub secret_access_key: Option<String>,
-        /// Session token
-        pub session_token: Option<String>,
-        /// AWS Region
-        pub region: Option<String>,
-        /// OSS or COS Endpoint
-        pub endpoint: Option<String>,
-        /// Allow HTTP (otherwise will always use https)
-        pub allow_http: Option<bool>,
     }
 }
